% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_map_state}
\alias{fars_map_state}
\title{Read in csv for  \code{year} and plot all the accidents in \code{state}
on a map.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer state number}

\item{year}{An integer year}
}
\value{
This function returns a list of tibbles
}
\description{
Read in csv for  \code{year} and plot all the accidents in \code{state}
on a map.
}
\note{
function will produce an error if there is no state data for
the \code{state} selected.

function will produce a warning if there are no accidents in
state, year combination in data
}
\examples{
fars_map_state(1, 2013)
}
