% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_read}
\alias{fars_read}
\title{Read csv}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string with the name of the csv to be loaded}
}
\value{
This function returns a dplyr tbl_df corresponding to the csv.
}
\description{
This is a function that reads in a csv names \code{filename} and
turns this into a dplyr tbl_df.
}
\note{
function will return an error if filename does not exist
}
\examples{
acc13 <- fars_read("accident_2013.csv.bz2")

}
