% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_read_years}
\alias{fars_read_years}
\title{Read in csv for eah year in \code{years} and returns the
months and year in that csv}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of year integers}
}
\value{
This function returns a list of tibbles
}
\description{
Read in csv for eah year in \code{years} and returns the
months and year in that csv
}
\note{
function will produce a warning if there is no filename
corresponding to year in years
}
\examples{
fars_read_years(c(2013,2014))
}
