% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{summarises the number of entries in the csv files for each
year and month in \code{years}}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of year integers}
}
\value{
This function returns a tibble that summarises each of the csv
files corresponding to the \code{years}
}
\description{
summarises the number of entries in the csv files for each
year and month in \code{years}
}
\examples{
fars_summarize_years(c(2013,2014))
}
