% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{make_filename}
\alias{make_filename}
\title{print filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer year (2013)}
}
\value{
This function returns a string that
corresponds to the name of a csv
}
\description{
This is a function takes a year integer \code{year} (eg 2013)
and returns the name of the csv
}
\examples{
make_filename(2013)
}
